+++ pvt/benchmarks/relayer.ts
@@ -14,10 +14,10 @@ const wordBytesSize = 32;
 
 async function main() {
   ({ vault } = await setupEnvironment());
-  relayerLibrary = await deploy('v2-standalone-utils/MockBaseRelayerLibrary', { args: [vault.address] });
+  relayerLibrary = await deploy('v2-standalone-utils/MockBaseRelayerLibrary', { args: [vault.address, ''] });
   const queryLibrary = await deploy('v2-standalone-utils/BatchRelayerQueryLibrary', { args: [vault.address] });
   relayer = await deploy('v2-standalone-utils/BalancerRelayer', {
-    args: [vault.address, relayerLibrary.address, queryLibrary.address],
+    args: [vault.address, relayerLibrary.address, queryLibrary.address, ''],
   });
   let totalGasUsed = bn(0);